requirements.txt
python-dotenv==1.1.0        # environment variable loader
Jinja2==3.1.6               # templating (used by FastAPI)
PyYAML==6.0.2               # YAML parsing
pydantic==2.11.4            # data validation
pydantic-ai==0.2.14         # pydantic extensions for AI
pandas==2.2.3
numpy==2.2.3
fastapi==0.115.12           # web framework
uvicorn[standard]==0.34.3   # ASGI server
openai==1.76.0              # OpenAI Python SDK
google-genai==1.19.0        # Google GenAI SDK
supabase==2.15.1            # Supabase client
PyMuPDF==1.25.3             # PDF parsing
chonkie[hub]==1.0.6         # miscellaneous utilities

litellm
matplotlib
seaborn